refactor(posv): Move UpdateM1 function to a new utils file for better organization and maintainability
